Another commit as ordered by robl

+++ Common/OpenGrid.Framework.Communications/RegionServerCommsManager.cs
@@ -50,7 +50,7 @@ namespace OpenGrid.Framework.Communications
         /// 
         /// </summary>
         /// <returns></returns>
-        public virtual bool InformNeighbourOfChildAgent( uint regionHandle, AgentCircuitData agentData) //should change from agentCircuitData
+        public virtual bool InformNeighbourOfChildAgent( ulong regionHandle, AgentCircuitData agentData) //should change from agentCircuitData
         {
             return false;
         }
